requirements.txt
urllib3
python-socketio[client]